Initial installable version of app.
It does nothing.
