+++ app/src/main/java/com/lima2017/neuralguide/api/INeuralGuideApi.java
@@ -1,4 +1,4 @@
-package uk.ac.cam.cl.lima2017.neuralguide.api;
+package com.lima2017.neuralguide.api;
 
 import android.media.Image;
 
@@ -6,9 +6,11 @@ import android.media.Image;
  * A generic representation of the Neural Guide API. This should match
  * exactly the endpoints provided by the JSON API.
  *
- * For Neural Guide WebAPI v0.1
+ * For Neural Guide WebAPI v0.1.
+ *
+ * @author Henry Thompson
+ * @version 1.0
  */
-
 public interface INeuralGuideApi {
     /**
      * Attempts to caption the image, and returns the result via a callback. The request may be
@@ -18,5 +20,5 @@ public interface INeuralGuideApi {
      * @param callback The callback that should be invoked when the attempt to caption the image
      *                 is completed.
      */
-    void tryCaptionImage(Image image, OnImageCaptionedListener callback);
+    void tryCaptionImage(final Image image, final OnImageCaptionedListener callback);
 }
