+++ app/src/main/java/com/lima2017/neuralguide/api/ImageCaptionResult.java
@@ -1,0 +1,52 @@
+package com.lima2017.neuralguide.api;
+
+import java.net.HttpURLConnection;
+
+/**
+ * A DTO for passing the results of an image captioning attempt to the user.
+ *
+ * For Neural Guide WebAPI v0.1
+ *
+ * @author Henry Thompson
+ * @version 1.0
+ */
+
+public class ImageCaptionResult {
+    /** The status code of the captioning attempt */
+    private final int _statusCode;
+
+    /** The actual caption returned by the server */
+    private final String _caption;
+
+    /**
+     * The result of querying the API to obtain a caption for the image.
+     * @param statusCode The status code associated with the result.
+     * @param caption The actual caption returned.
+     */
+    public ImageCaptionResult(final int statusCode, final String caption) {
+        _statusCode = statusCode;
+        _caption = caption;
+    }
+
+    /**
+     * @return <code>true</code> if and only if the API successfully generated a caption for the
+     * image, even if it is not confident about the result.
+     */
+    public boolean success() {
+        return _statusCode == HttpURLConnection.HTTP_OK;
+    }
+
+    /**
+     * @return The HTTP status code returned by the API.
+     */
+    public int getStatusCode() {
+        return _statusCode;
+    }
+
+    /**
+     * @return The caption generated by the API.
+     */
+    public String getCaption() {
+        return _caption;
+    }
+}
