+++ app/src/main/java/com/lima2017/neuralguide/api/demo/NeuralGuideApiDemoMock.java
@@ -1,13 +1,19 @@
-package uk.ac.cam.cl.lima2017.neuralguide.api.demo;
+package com.lima2017.neuralguide.api.demo;
 
 import android.media.Image;
 
 import java.net.HttpURLConnection;
 
-import uk.ac.cam.cl.lima2017.neuralguide.api.INeuralGuideApi;
-import uk.ac.cam.cl.lima2017.neuralguide.api.OnImageCaptionedListener;
-import uk.ac.cam.cl.lima2017.neuralguide.api.ImageCaptionResult;
+import com.lima2017.neuralguide.api.INeuralGuideApi;
+import com.lima2017.neuralguide.api.OnImageCaptionedListener;
+import com.lima2017.neuralguide.api.ImageCaptionResult;
 
+/**
+ * Mocks the API by returning a hardcoded string value.
+ *
+ * @author Henry Thompson
+ * @version 1.0
+ */
 public class NeuralGuideApiDemoMock implements INeuralGuideApi {
     @Override
     public void tryCaptionImage(Image image, OnImageCaptionedListener callback) {
