--- app/src/main/java/uk/ac/cam/cl/lima2017/neuralguide/NeuralGuideFragment.java
@@ -1,20 +1,0 @@
-package uk.ac.cam.cl.lima2017.neuralguide;
-
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-/**
- * Created by henry on 1/31/17.
- */
-
-public class NeuralGuideFragment extends Fragment {
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_neural_guide, container, false);
-    }
-}
