--- app/src/main/java/uk/ac/cam/cl/lima2017/neuralguide/api/web/WebApiConfig.java
@@ -1,6 +1,0 @@
-package uk.ac.cam.cl.lima2017.neuralguide.api.web;
-
-public class WebApiConfig {
-    public static final String VERSION = "0.1";
-    public static final String BASE_URL = "neuralguide.com";
-}
