+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,6 @@
 <resources>
     <string name="app_name">Neural Guide</string>
 
-    <string name="dummy_button">Dummy Button</string>
-    <string name="dummy_content">DUMMY\nCONTENT</string>
+    <string name="camera_tap_or_speak_prompt">Tap or say \"What is this?\"</string>
+    <string name="camera_feedback_placeholder_text">I think this is a dog</string>
 </resources>
