+++ app/src/test/java/com/lima2017/neuralguide/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package uk.ac.cam.cl.lima2017.neuralguide;
+package com.lima2017.neuralguide;
 
 import org.junit.Test;
